+++ R/MiscFuns.R
@@ -2949,21 +2949,28 @@ bin = function(x, bin){
 #'
 #' The definitions of the macro variables will replace in verbatim the macro variables. Therefore, you can include multi-part formulas if you wish but then beware of the order of the macros variable in the formula. For example, using the \code{airquality} data, say you want to set as controls the variable \code{Temp} and \code{Day} fixed-effects, you can do \code{setFixest_fml(..ctrl = ~Temp | Day)}, but then \code{feols(Ozone ~ Wind + ..ctrl, airquality)} will be quite different from \code{feols(Ozone ~ ..ctrl + Wind, airquality)}, so beware!
 #'
-#' @section Dotted square brackets in formulas:
+#' @section Dot square bracket operator in formulas:
 #'
-#' In the formula, square brackets can i) create manifold variables at once, or ii) capture values from the current environment and put them verbatim in the formula. Note that the square bracket must be preceded by a dot.
+#' In a formula, the dot square bracket (DSB) operator can: i) create manifold variables at once, or ii) capture values from the current environment and put them verbatim in the formula.
 #'
 #' Say you want to include the variables \code{x1} to \code{x3} in your formula. You can use \code{xpd(y ~ x.[1:3])} and you'll get \code{y ~ x1 + x2 + x3}.
-#'
-#' You can use square brackets within variable names but they must be in character form in a macro variable: \code{xpd(y ~ ..x, ..x = "x.[1:2]_sq")} will create \code{y ~ x1_sq + x2_sq}. Otherwise there would be parsing issues.
 #'
 #' To summon values from the environment, simply put the variable in square brackets. For example: \code{for(i in 1:3) xpd(y.[i] ~ x)} will create the formulas \code{y1 ~ x} to \code{y3 ~ x} depending on the value of \code{i}.
 #'
 #' You can include a full variable from the environment in the same way: \code{for(y in c("a", "b")) xpd(.[y] ~ x)} will create the two formulas \code{a ~ x} and \code{b ~ x}.
 #'
-#' In all \code{fixest} estimations, this special parsing is enabled, so you don't need to use \code{xpd} (except for square brackets within a variable name).
+#' The DSB can even be used within variable names, but then the variable must be nested in character form. For example \code{y ~ .["x.[1:2]_sq"]} will create \code{y ~ x1_sq +  x2_sq}. Using the character form is important to avoid a formula parsing error.
+#'
+#' In all \code{fixest} estimations, this special parsing is enabled, so you don't need to use \code{xpd}.
 #'
 #' Limitations: the use of multiple square brackets within a single variable is not implemented. For example, the following will not work \code{xpd(y ~ ..x, ..x = x.[1:3]_.[1:3])}.
+#'
+#' @section Regular expressions:
+#'
+#' You can catch several variable names at once by using regular expressions. To use regular expressions, you need to enclose it in the dot-dot function: \code{..("regex")}. For example, \code{..("Sepal")} will catch both the variables \code{Sepal.Length} and \code{Sepal.Width} from the \code{iris} data set. In a \code{fixest} estimation, the variables names from which the regex will be applied come from the data set. If you use \code{xpd}, you need to provide either a data set or a vector of names in the argument \code{data}.
+#'
+#' Note that the dot square bracket operator (DSB, see before) is applied before the regular expression is evaluated. This means that \code{..("x.[3:4]_sq")} will lead, after evaluation of the DSB, to \code{..("x3_sq|x4_sq")}. It is a handy way to insert range of numbers in a regular expression.
+#'
 #'
 #' @return
 #' It returns a formula where all macros have been expanded.
@@ -3083,11 +3090,17 @@ bin = function(x, bin){
 #' xpd(y ~ x.[var])
 #'
 #' # ... the variables can be multiple
-#' vars = letters[1:5]
+#' vars = LETTERS[1:3]
 #' xpd(y ~ x.[vars])
 #'
-#' # You can have "complex" variable names but must be in character form
-#' xpd(y ~ ..x, ..x = "x.[vars]_sq")
+#' # You can have "complex" variable names but they must be nested in character form
+#' xpd(y ~ .["x.[vars]_sq"])
+#'
+#' # DSB can be used within regular expressions
+#' re = c("GNP", "Pop")
+#' xpd(Unemployed ~ ..(".[re]"), data = longley)
+#'
+#' # => equivalent to ..("GNP|Pop")
 #'
 #'
 xpd = function(fml, ..., lhs, rhs, data = NULL){
@@ -3182,12 +3195,14 @@ xpd = function(fml, ..., lhs, rhs, data = NULL){
             res = fml_dp_split
             for(i in 2:length(res)){
                 re = gsub('"\\).*', "", res[i])
-                vars = grep(re, data, value = TRUE)
+                re_width = nchar(re)
+                re = dot_square_bracket(re, frame, regex = TRUE)
+                vars = grep(re, data, value = TRUE, perl = TRUE)
                 if(length(vars) == 0){
                     vars = "1"
                 }
 
-                res[i] = paste0(paste(vars, collapse = "+"), substr(res[i], nchar(re) + 3, nchar(res[i])))
+                res[i] = paste0(paste(vars, collapse = "+"), substr(res[i], re_width + 3, nchar(res[i])))
             }
 
             fml = as.formula(paste(res, collapse = ""), frame)
@@ -3197,11 +3212,11 @@ xpd = function(fml, ..., lhs, rhs, data = NULL){
     if("[" %in% all.vars(fml, functions = TRUE)){
         fml_txt = deparse_long(fml)
         if(grepl(".[", fml_txt, fixed = TRUE)){
-            fml_txt = charget(fml_txt, frame)
+            fml_txt = dot_square_bracket(fml_txt, frame)
 
             # level 1 recursivity
             if(grepl(".[", fml_txt, fixed = TRUE)){
-                fml_txt = charget(fml_txt, frame)
+                fml_txt = dot_square_bracket(fml_txt, frame)
             }
 
             fml = as.formula(fml_txt, frame)
@@ -3983,7 +3998,7 @@ value2stringCall = function(value_raw, call = FALSE, check = FALSE, frame = NULL
                 # We need to check that it leads to a valid formula => otherwise problems later
 
                 for(i in seq_along(value_raw)){
-                    value_raw[i] = charget(value_raw[i], frame)
+                    value_raw[i] = dot_square_bracket(value_raw[i], frame)
                 }
 
                 value_raw = paste(value_raw, collapse = " + ")
@@ -4007,21 +4022,52 @@ value2stringCall = function(value_raw, call = FALSE, check = FALSE, frame = NULL
     }
 
     if(check && "[" %in% all.vars(res, functions = TRUE)){
-        res_txt = charget(deparse_long(res), frame)
+        res_txt = dot_square_bracket(deparse_long(res), frame)
         res = as.formula(res_txt)
     }
 
     res
 }
 
-charget = function(x, frame = .GlobalEnv){
-    # tansforms "x.[i]" into x1 if i==1
+dot_square_bracket = function(x, frame = .GlobalEnv, regex = FALSE){
+    # transforms "x.[i]" into x1 if i==1
     # z = "XX" ; x = ".[z] + x.[1:5] + y.[1:2]_t"
     # x = "x.[a] | fe[b] + m.[o]"
 
     if(!grepl(".[", x, fixed = TRUE)) return(x)
 
     x_split_open = strsplit(x, ".[", fixed = TRUE)[[1]]
+
+    # nesting: a ~ .["x.[1:2]_sq"] + b
+    is_nested = any(grepl("^\"", x_split_open))
+    if(is_nested){
+
+        i_open_quote = which(grepl("^\"", x_split_open))
+        i_close_quote = which(grepl("\"\\]", x_split_open))
+
+        x_split_new = x_split_open[1]
+        for(i in i_open_quote){
+            j = i_close_quote[i_close_quote >= i]
+
+            xi_new_all = x_split_open[i:j]
+            xi_new_all = gsub("\\]", "__close__", xi_new_all)
+
+            xi_new = paste0(xi_new_all, collapse = "__open__")
+            xi_new = gsub("\"__close__", "\"]", xi_new)
+
+            x_split_new[[length(x_split_new) + 1]] = xi_new
+        }
+
+        n = length(x_split_open)
+        if(j < n){
+            for(i in (j+1):n){
+                x_split_new[[length(x_split_new) + 1]] = x_split_open[i]
+            }
+        }
+
+        x_split_open = x_split_new
+    }
+
     x_split_close_left = gsub("^([^\\[]+)\\].*", "\\1", x_split_open[-1])
     x_split_close_right = substr(x_split_open[-1], nchar(x_split_close_left) + 2, nchar(x_split_open[-1]))
 
@@ -4031,6 +4077,11 @@ charget = function(x, frame = .GlobalEnv){
 
     for(i in 2:n){
         x_split[[i]] = if(i %% 2 == 0) x_split_close_left[i %/% 2] else x_split_close_right[i %/% 2]
+    }
+
+    if(is_nested){
+        x_split = gsub("__open__", ".[", x_split)
+        x_split = gsub("__close__", "]", x_split)
     }
 
     res = as.list(x_split)
@@ -4044,6 +4095,7 @@ charget = function(x, frame = .GlobalEnv){
         res = paste(res, collapse = "")
 
     } else {
+        # first value is NEVER a vector ("" is added automatically in split)
         res_txt = res[[1]]
         i = 2
         while(i <= n){
@@ -4051,6 +4103,11 @@ charget = function(x, frame = .GlobalEnv){
             if(length(res[[i]]) == 1){
                 res_txt = paste0(res_txt, res[[i]])
                 i = i + 1
+
+            } else if(regex) {
+                after = if(i != n) res[[i + 1]] else ""
+                res_txt = paste0(res_txt, res[[i]], after, collapse = "|")
+                i = i + 2
 
             } else {
                 before_no_var = gsub("[[:alnum:]_\\.]+$", "", res_txt)
@@ -4071,6 +4128,10 @@ charget = function(x, frame = .GlobalEnv){
         }
 
         res = res_txt
+    }
+
+    if(grepl(".[", res, fixed = TRUE)){
+        res = dot_square_bracket(res, frame, regex)
     }
 
     res